refactor(desktop): extract realtime hub policies (phase 2, move-only)
