+++ desktop/macos/Desktop/Sources/FloatingControlBar/RealtimeHubController.swift
@@ -1,6 +1,5 @@
 import AppKit
 import CoreGraphics
-import CryptoKit
 import Foundation
 import OmiSupport
 
@@ -17,1016 +16,6 @@ import OmiSupport
 //
 // Provider tool proposals are untrusted until the kernel resolves the canonical
 // route and authorizes the active run/attempt capability.
-
-/// Safe, non-sensitive classification for realtime WebSocket teardown messages.
-///
-/// Gemini can idle-close warm sessions with WebSocket 1008 after the socket has
-/// lived for a while, and OpenAI has a known maximum-session-duration close.
-/// Both expected lifecycle paths should re-warm quietly rather than page Sentry
-/// as production errors. Fast 1008 closes are different: they usually mean
-/// provider policy/auth/config rejection and should still be reported, but with
-/// a stable category instead of raw provider text.
-enum RealtimeHubCloseCategory: String {
-  case expectedIdleTeardown = "expected_idle_teardown"
-  case expectedSessionRotation = "expected_session_rotation"
-  case providerAuthFailed = "provider_auth_failed"
-  case providerQuotaExceeded = "provider_quota_exceeded"
-  case providerPolicyCloseFast = "provider_policy_close_fast"
-  case providerTransient = "provider_transient"
-}
-
-/// The controller owns transport replacement, while the voice-turn reducer owns
-/// any active logical turn. Keeping this plan typed prevents provider-close text
-/// from leaking into lifecycle decisions outside the classifier boundary.
-enum RealtimeHubSessionRotationPlan: Equatable {
-  case rewarmIdleTransport
-  case terminateActiveTurnAndRewarm
-}
-
-enum RealtimeHubCloseClassifier {
-  static let idleTeardownThreshold: TimeInterval = 60
-
-  static func category(
-    message: String,
-    aliveFor: TimeInterval,
-    hasActiveTurn: Bool = false,
-    provider: RealtimeHubProvider = .openai
-  ) -> RealtimeHubCloseCategory? {
-    let lower = message.lowercased()
-    if provider == .openai,
-      lower.contains("your session hit the maximum duration")
-      && lower.contains("60 minutes")
-    {
-      return .expectedSessionRotation
-    }
-    guard lower.contains("websocket closed (1008)") else { return nil }
-    if CredentialHealthManager.classifyProviderClose(
-      message: message,
-      provider: provider) == .providerQuotaExceeded(provider: provider)
-    {
-      return .providerQuotaExceeded
-    }
-    if CredentialHealthManager.classifyProviderClose(
-      message: message,
-      provider: provider) == .providerAuthFailed(provider: provider, mode: .byok)
-    {
-      return .providerAuthFailed
-    }
-    if !hasActiveTurn && aliveFor >= idleTeardownThreshold { return .expectedIdleTeardown }
-    return .providerPolicyCloseFast
-  }
-
-  static func sessionRotationPlan(
-    for category: RealtimeHubCloseCategory?,
-    hasActiveTurn: Bool
-  ) -> RealtimeHubSessionRotationPlan? {
-    guard category == .expectedSessionRotation else { return nil }
-    return hasActiveTurn ? .terminateActiveTurnAndRewarm : .rewarmIdleTransport
-  }
-
-  static func isExpectedLifecycleClose(_ category: RealtimeHubCloseCategory?) -> Bool {
-    category == .expectedIdleTeardown || category == .expectedSessionRotation
-  }
-
-  static func shouldReportToSentry(_ category: RealtimeHubCloseCategory?) -> Bool {
-    !isExpectedLifecycleClose(category)
-  }
-}
-
-enum RealtimeHubCommitResult: Equatable {
-  case accepted
-  case deferredForReplacement
-  case deferredForReconnect
-  case alreadyOwned
-  case rejectedNoSession
-}
-
-enum RealtimeHubCommitOwnershipPolicy {
-  static func isAlreadyOwned(turn: VoiceTurn?, requestedTurnID: VoiceTurnID) -> Bool {
-    guard let turn, turn.id == requestedTurnID, turn.phase == .awaitingResponse,
-      turn.hubCommitPending
-    else { return false }
-    if case .hub = turn.route { return true }
-    return false
-  }
-}
-
-struct RealtimeProviderToolResult: Equatable {
-  let output: String
-  let originalByteCount: Int
-  let wasOversized: Bool
-}
-
-enum RealtimeProviderToolResultPolicy {
-  static let maximumByteCount = 48 * 1024
-
-  /// The sole model-visible result boundary for both realtime providers.
-  /// Provider-specific transport only happens after this method returns a
-  /// canonical envelope, including rejected/authorization error paths.
-  static func prepare(
-    provider: RealtimeHubProvider = .openai,
-    name: String,
-    output: String
-  ) -> RealtimeProviderToolResult {
-    let originalByteCount = output.utf8.count
-    var providerOutput = output
-    if name == HubTool.spawnAgent.rawValue,
-      let payload = try? JSONSerialization.jsonObject(with: Data(output.utf8)) as? [String: Any],
-      let providerResult = payload["providerResult"] as? [String: Any],
-      isCanonicalToolResultEnvelope(providerResult["toolResultEnvelope"]),
-      let data = try? JSONSerialization.data(withJSONObject: providerResult, options: [.sortedKeys]),
-      let compact = String(data: data, encoding: .utf8)
-    {
-      providerOutput = compact
-    }
-    guard providerOutput.utf8.count <= maximumByteCount else {
-      return oversizedFailure(
-        provider: provider,
-        name: name,
-        originalOutput: output,
-        originalByteCount: originalByteCount)
-    }
-    let finalized = finalize(provider: provider, name: name, output: providerOutput)
-    guard finalized.utf8.count <= maximumByteCount else {
-      return oversizedFailure(
-        provider: provider,
-        name: name,
-        originalOutput: output,
-        originalByteCount: originalByteCount)
-    }
-    return RealtimeProviderToolResult(
-      output: finalized,
-      originalByteCount: originalByteCount,
-      wasOversized: false)
-  }
-
-  static func rejectedOutput(
-    code: String,
-    message: String,
-    preservingCanonicalEnvelopeFrom sourceOutput: String? = nil
-  ) -> String {
-    var payload: [String: Any] = [
-      "ok": false,
-      "error": ["code": String(code.prefix(128)), "message": String(message.prefix(512))],
-    ]
-    // Node owns the capability tuple. A Swift classification may refine the
-    // user-facing error, but it must never replace a valid external-run
-    // envelope with synthetic realtime provenance.
-    if let sourceOutput,
-      let sourcePayload = try? JSONSerialization.jsonObject(with: Data(sourceOutput.utf8)) as? [String: Any],
-      isCanonicalToolResultEnvelope(sourcePayload["toolResultEnvelope"])
-    {
-      payload["toolResultEnvelope"] = sourcePayload["toolResultEnvelope"]
-    }
-    let data = try? JSONSerialization.data(withJSONObject: payload, options: [.sortedKeys])
-    return data.flatMap { String(data: $0, encoding: .utf8) }
-      ?? #"{"ok":false,"error":{"code":"realtime_tool_rejected","message":"The tool could not be completed."}}"#
-  }
-
-  private static func oversizedFailure(
-    provider: RealtimeHubProvider,
-    name: String,
-    originalOutput: String,
-    originalByteCount: Int
-  ) -> RealtimeProviderToolResult {
-      let error: [String: Any] = [
-        "code": "tool_result_too_large",
-        "message": "The tool returned too much detail. Retry with narrower filters.",
-        "tool": String(name.prefix(128)),
-        "originalBytes": originalByteCount,
-      ]
-      let payload: [String: Any] = [
-        "ok": false,
-        "error": error,
-        "toolResultEnvelope": providerFailureEnvelope(provider: provider, name: name, error: error, originalOutput: originalOutput),
-      ]
-      let data = try? JSONSerialization.data(withJSONObject: payload, options: [.sortedKeys])
-      let bounded = data.flatMap { String(data: $0, encoding: .utf8) }
-        ?? #"{"ok":false,"error":{"code":"tool_result_too_large"}}"#
-      return RealtimeProviderToolResult(
-        output: bounded,
-        originalByteCount: originalByteCount,
-        wasOversized: true)
-  }
-
-  private static func finalize(provider: RealtimeHubProvider, name: String, output: String) -> String {
-    if parsedCanonicalToolResultEnvelope(from: output) != nil { return output }
-    guard var payload = try? JSONSerialization.jsonObject(with: Data(output.utf8)) as? [String: Any] else {
-      return rejectedOutputEnvelope(
-        provider: provider,
-        name: name,
-        code: "realtime_tool_result_unstructured",
-        message: "The tool returned an invalid response.")
-    }
-    let status = payload["ok"] as? Bool == true ? "succeeded" : "failed"
-    let payloadBytes = (try? JSONSerialization.data(withJSONObject: payload, options: [.sortedKeys]))?.count ?? 0
-    payload["toolResultEnvelope"] = [
-      "version": 1,
-      "status": status,
-      "truncated": false,
-      "originalBytes": payloadBytes,
-      "projectedBytes": payloadBytes,
-      "fullOutputRef": NSNull(),
-      "provenance": [
-        "invocationId": "realtime-\(provider.rawValue)-\(String(name.prefix(64)))",
-        "runId": "unknown",
-        "attemptId": "unknown",
-        "toolName": String(name.prefix(128)),
-      ],
-    ]
-    let data = try? JSONSerialization.data(withJSONObject: payload, options: [.sortedKeys])
-    return data.flatMap { String(data: $0, encoding: .utf8) }
-      ?? rejectedOutputEnvelope(
-        provider: provider,
-        name: name,
-        code: "realtime_tool_result_encoding_failed",
-        message: "The tool response could not be prepared.")
-  }
-
-  private static func rejectedOutputEnvelope(
-    provider: RealtimeHubProvider,
-    name: String,
-    code: String,
-    message: String
-  ) -> String {
-    let error: [String: Any] = ["code": String(code.prefix(128)), "message": String(message.prefix(512))]
-    let payload: [String: Any] = [
-      "ok": false,
-      "error": error,
-      "toolResultEnvelope": providerFailureEnvelope(
-        provider: provider,
-        name: name,
-        error: error,
-        originalOutput: ""),
-    ]
-    let data = try? JSONSerialization.data(withJSONObject: payload, options: [.sortedKeys])
-    return data.flatMap { String(data: $0, encoding: .utf8) }
-      ?? #"{"ok":false}"#
-  }
-
-  /// The Node bridge owns artifacts. Swift never replaces a canonical reference
-  /// with a provider-only payload; it either forwards that envelope or returns
-  /// a typed provider-budget error which preserves an existing artifact ref.
-  private static func providerFailureEnvelope(
-    provider: RealtimeHubProvider,
-    name: String,
-    error: [String: Any],
-    originalOutput: String
-  ) -> [String: Any] {
-    let sourceEnvelope = parsedCanonicalToolResultEnvelope(from: originalOutput)
-    let errorBytes = (try? JSONSerialization.data(withJSONObject: error, options: [.sortedKeys]))?.count ?? 0
-    let sourceOriginalBytes = sourceEnvelope?["originalBytes"] as? Int
-    let sourceRef = sourceEnvelope?["fullOutputRef"] as? String
-    let preservesArtifact = sourceRef != nil && (sourceOriginalBytes ?? 0) > errorBytes
-    let sourceProvenance = sourceEnvelope?["provenance"] as? [String: Any]
-    return [
-      "version": 1,
-      "status": "failed",
-      "truncated": preservesArtifact,
-      "originalBytes": preservesArtifact ? sourceOriginalBytes! : errorBytes,
-      "projectedBytes": errorBytes,
-      "fullOutputRef": preservesArtifact ? sourceRef! : NSNull(),
-      "provenance": sourceProvenance ?? [
-        "invocationId": "realtime-\(provider.rawValue)-provider-budget-\(String(name.prefix(64)))",
-        "runId": "unknown",
-        "attemptId": "unknown",
-        "toolName": String(name.prefix(128)),
-      ],
-    ]
-  }
-
-  private static func parsedCanonicalToolResultEnvelope(from output: String) -> [String: Any]? {
-    guard let payload = try? JSONSerialization.jsonObject(with: Data(output.utf8)) as? [String: Any] else {
-      return nil
-    }
-    if let envelope = payload["toolResultEnvelope"] as? [String: Any], isCanonicalToolResultEnvelope(envelope) {
-      return envelope
-    }
-    if let providerResult = payload["providerResult"] as? [String: Any],
-      let envelope = providerResult["toolResultEnvelope"] as? [String: Any],
-      isCanonicalToolResultEnvelope(envelope)
-    {
-      return envelope
-    }
-    return nil
-  }
-
-  private static func isCanonicalToolResultEnvelope(_ value: Any?) -> Bool {
-    guard let envelope = value as? [String: Any],
-      envelope["version"] as? Int == 1,
-      ["succeeded", "failed", "cancelled"].contains(envelope["status"] as? String ?? ""),
-      let originalBytes = envelope["originalBytes"] as? Int,
-      let projectedBytes = envelope["projectedBytes"] as? Int,
-      let truncated = envelope["truncated"] as? Bool,
-      originalBytes >= projectedBytes,
-      truncated == (projectedBytes < originalBytes),
-      let provenance = envelope["provenance"] as? [String: Any]
-    else { return false }
-    guard ["invocationId", "runId", "attemptId", "toolName"].allSatisfy({
-      (provenance[$0] as? String)?.isEmpty == false
-    }) else { return false }
-    if projectedBytes < originalBytes { return envelope["fullOutputRef"] as? String != nil }
-    return envelope["fullOutputRef"] is NSNull || envelope["fullOutputRef"] == nil
-  }
-}
-
-struct RealtimeHubLifecycleSnapshot: Equatable {
-  let capturingInput: Bool
-  let providerActive: Bool
-  let playbackActive: Bool
-  let pendingToolCount: Int
-  let coordinatorTurnActive: Bool
-  let minting: Bool
-}
-
-enum RealtimeHubLifecyclePolicy {
-  static func canReplaceSession(_ snapshot: RealtimeHubLifecycleSnapshot) -> Bool {
-    !snapshot.capturingInput
-      && !snapshot.providerActive
-      && !snapshot.playbackActive
-      && snapshot.pendingToolCount == 0
-      && !snapshot.coordinatorTurnActive
-      && !snapshot.minting
-  }
-
-  static func canStartGeneralWarmSession(replacementPending: Bool) -> Bool {
-    !replacementPending
-  }
-
-  static func shouldResumeCanceledTurnRefresh(
-    fenceTurnID: VoiceTurnID?,
-    terminalTurnID: VoiceTurnID
-  ) -> Bool {
-    fenceTurnID != terminalTurnID
-  }
-}
-
-/// Rebuilds hub decisions from kernel-owned continuity facts after relaunch.
-/// Hub maps die with the process; stable turn IDs already in the kernel journal
-/// (or its typed voice-context snapshot) are the durable ownership signal.
-enum RealtimeHubContinuityRestore {
-  static func kernelOwnsExchange(
-    continuityKey: String,
-    kernelTurnIDs: Set<String>
-  ) -> Bool {
-    guard !continuityKey.isEmpty else { return false }
-    return !kernelTurnIDs.isDisjoint(
-      with: KernelTurnProjection.stableTurnIDs(continuityKey: continuityKey))
-  }
-}
-
-/// Selects the sole journal owner after realtime `spawn_agent` admission. The
-/// spawn RPC atomically records the canonical exchange; provider turn_done may
-/// refresh that projection, but it must never submit a second mutation for the
-/// same continuity identity. After relaunch, `kernelOwnsExchange` restores that
-/// same authority when hub-local spawn-receipt maps are empty.
-@MainActor
-enum RealtimeTurnJournalAuthority {
-  static func persist(
-    turnOwnerID: String,
-    acceptedSpawnOwnerID: String?,
-    kernelOwnsExchange: Bool = false,
-    refreshAcceptedSpawn: @escaping @MainActor () async -> Bool,
-    recordProviderExchange: @escaping @MainActor () async -> Bool
-  ) async -> Bool {
-    if acceptedSpawnOwnerID == turnOwnerID || kernelOwnsExchange {
-      return await refreshAcceptedSpawn()
-    }
-    return await recordProviderExchange()
-  }
-}
-
-/// A single kernel write is an obligation of one stable continuity key.  The
-/// controller may have several turns in flight while a barge-in replaces a
-/// physical session, so completion of B must never supersede A's receipt.
-struct RealtimeTurnPersistenceReceipt: Equatable {
-  let continuityKey: String
-  let accepted: Bool
-}
-
-@MainActor
-final class RealtimeTurnPersistenceLedger {
-  private struct Obligation {
-    let id: UUID
-    let task: Task<Bool, Never>
-    var retainingReceipt: Bool
-  }
-
-  private var obligations: [String: Obligation] = [:]
-  private var receipts: [String: RealtimeTurnPersistenceReceipt] = [:]
-  private(set) var generation: UInt64 = 0
-
-  var pendingContinuityKeys: Set<String> {
-    Set(obligations.keys)
-  }
-
-  /// Idempotent and widen-never-narrow: a continuity key already in flight
-  /// returns its existing task unchanged rather than starting a second kernel
-  /// write. But retention is a live property of that obligation, not a value
-  /// frozen into the first caller's closure — a later call that additionally
-  /// needs a receipt (`retainingReceipt: true`) widens the same obligation in
-  /// place, and the completion below re-reads the current flag instead of the
-  /// value captured at creation, so the widening call is never lost to the
-  /// first caller's capture.
-  @discardableResult
-  func enqueue(
-    continuityKey: String,
-    retainingReceipt: Bool,
-    _ operation: @escaping @MainActor () async -> Bool
-  ) -> Task<Bool, Never> {
-    if var existing = obligations[continuityKey] {
-      if retainingReceipt, !existing.retainingReceipt {
-        existing.retainingReceipt = true
-        obligations[continuityKey] = existing
-      }
-      return existing.task
-    }
-
-    receipts.removeValue(forKey: continuityKey)
-    generation &+= 1
-    let obligationID = UUID()
-    let task = Task { @MainActor [weak self] in
-      let accepted = await operation()
-      guard let self,
-        self.obligations[continuityKey]?.id == obligationID
-      else {
-        // The kernel result remains truthful to this caller even if an owner
-        // transition or a newer obligation has retired this local ledger entry.
-        return accepted
-      }
-      // Re-read retention live rather than the value this closure captured at
-      // creation, so a widening `enqueue(retainingReceipt: true)` observed
-      // after this task started still gets its receipt.
-      let shouldRetain = self.obligations[continuityKey]?.retainingReceipt ?? retainingReceipt
-      self.obligations.removeValue(forKey: continuityKey)
-      // An external authority (e.g. an atomically-accepted spawn) may have
-      // already recorded this key's receipt while this obligation was still
-      // pending via `recordAcceptedReceipt`. That acceptance stays
-      // authoritative — this obligation's own completion must never overwrite it.
-      if shouldRetain, self.receipts[continuityKey] == nil {
-        self.receipts[continuityKey] = RealtimeTurnPersistenceReceipt(
-          continuityKey: continuityKey,
-          accepted: accepted)
-      }
-      return accepted
-    }
-    obligations[continuityKey] = Obligation(
-      id: obligationID, task: task, retainingReceipt: retainingReceipt)
-    return task
-  }
-
-  /// Awaits the exact obligation for this logical turn and consumes only its
-  /// receipt.  A concurrent B write cannot alter A's outcome.
-  func consumeReceipt(for continuityKey: String) async -> RealtimeTurnPersistenceReceipt? {
-    if let obligation = obligations[continuityKey] {
-      _ = await obligation.task.value
-    }
-    return receipts.removeValue(forKey: continuityKey)
-  }
-
-  func receipt(for continuityKey: String) -> RealtimeTurnPersistenceReceipt? {
-    receipts[continuityKey]
-  }
-
-  /// Use only for a kernel mutation that was atomically accepted by a different
-  /// authority (currently `spawn_agent`).  It gives that exact continuity key a
-  /// finalization receipt without inventing a second provider-exchange write.
-  /// Idempotent: recording the same key's acceptance twice (e.g. a duplicate
-  /// external acceptance callback) writes the same `accepted: true` payload
-  /// both times, so it is safe to call more than once without an assertion.
-  func recordAcceptedReceipt(for continuityKey: String) {
-    receipts[continuityKey] = RealtimeTurnPersistenceReceipt(
-      continuityKey: continuityKey,
-      accepted: true)
-  }
-
-  /// Repeatedly observes pending obligations because a new turn may enqueue a
-  /// write while an earlier write is suspended.  It never serializes unrelated
-  /// continuity keys.
-  func awaitPendingObligations() async {
-    while !Task.isCancelled {
-      let pending = obligations.values.map(\.task)
-      guard !pending.isEmpty else { return }
-      for task in pending {
-        _ = await task.value
-      }
-    }
-  }
-
-  /// Owner replacement revokes every local waiter.  An ignored cancellation can
-  /// still complete its kernel RPC, but its old obligation is forbidden from
-  /// removing or writing over any newer key's state.
-  func cancelAll() {
-    let pending = obligations.values.map(\.task)
-    obligations.removeAll()
-    receipts.removeAll()
-    generation &+= 1
-    for task in pending { task.cancel() }
-  }
-}
-
-/// Immutable account identity attached to a realtime socket, its context, and
-/// every token mint that may create or replace it. `signedOut` is a real scope,
-/// distinct from the absence of a session, so a later sign-in cannot inherit a
-/// socket that was warmed from signed-out/BYOK state.
-enum RealtimeHubOwnerScope: Equatable, Sendable {
-  case authenticated(String)
-  case signedOut
-
-  static func capture(currentOwnerID: String?) -> RealtimeHubOwnerScope {
-    guard let ownerID = currentOwnerID?.trimmingCharacters(in: .whitespacesAndNewlines),
-      !ownerID.isEmpty
-    else { return .signedOut }
-    return .authenticated(ownerID)
-  }
-
-  var authenticatedOwnerID: String? {
-    guard case .authenticated(let ownerID) = self else { return nil }
-    return ownerID
-  }
-
-  func isCurrent(currentOwnerID: String?) -> Bool {
-    self == Self.capture(currentOwnerID: currentOwnerID)
-  }
-}
-
-#if DEBUG
-struct RealtimeHubOwnerBoundarySnapshot: Equatable {
-  let hasPhysicalSession: Bool
-  let physicalOwnerID: String?
-  let prefetchedOwnerID: String?
-  let prefetchedContextIsEmpty: Bool
-  let hasPendingOwnerWork: Bool
-  let hubConnected: Bool
-  let turnAudioByteCount: Int
-}
-
-/// Exact non-production capability for the hermetic local-profile transport.
-/// A process-wide "test mode" boolean is not enough: the authority is bound to
-/// one physical session and one immutable owner scope, so a replaced socket or
-/// an owner transition cannot inherit the provider-warm bypass.
-struct RealtimeLocalProfileTransportAuthority: Equatable {
-  let sourceID: ObjectIdentifier
-  let ownerScope: RealtimeHubOwnerScope
-  let authorizationSnapshot: RuntimeOwnerAuthorizationSnapshot
-
-  func accepts(
-    sourceID candidateSourceID: ObjectIdentifier?,
-    currentOwnerID: String?,
-    localProfileEnabled: Bool,
-    authorizationIsCurrent: Bool
-  ) -> Bool {
-    localProfileEnabled
-      && candidateSourceID == sourceID
-      && ownerScope.isCurrent(currentOwnerID: currentOwnerID)
-      && authorizationIsCurrent
-  }
-}
-#endif
-
-/// Shared owner policy used by warm reuse, delayed mint completion, and
-/// barge-in replacement. Keeping these three paths on one decision surface
-/// prevents a future reconnect path from accidentally weakening the fence.
-enum RealtimeHubOwnerFence {
-  static func canReuseWarmSession(
-    sessionOwner: RealtimeHubOwnerScope?,
-    currentOwnerID: String?
-  ) -> Bool {
-    guard let sessionOwner else { return false }
-    return sessionOwner.isCurrent(currentOwnerID: currentOwnerID)
-  }
-
-  static func acceptsMintCompletion(
-    mintOwner: RealtimeHubOwnerScope,
-    currentOwnerID: String?
-  ) -> Bool {
-    mintOwner.isCurrent(currentOwnerID: currentOwnerID)
-  }
-
-  static func acceptsBargeInReplacement(
-    sessionOwner: RealtimeHubOwnerScope?,
-    replacementOwner: RealtimeHubOwnerScope?,
-    currentOwnerID: String?
-  ) -> Bool {
-    guard let sessionOwner, let replacementOwner, sessionOwner == replacementOwner else {
-      return false
-    }
-    return replacementOwner.isCurrent(currentOwnerID: currentOwnerID)
-  }
-}
-
-enum RealtimeNativeAudioScheduleFailureAction: Equatable {
-  case keepTextFallback
-  case failTurnAfterPartialPlayback
-
-  static func decide(playbackAlreadyStarted: Bool) -> RealtimeNativeAudioScheduleFailureAction {
-    playbackAlreadyStarted ? .failTurnAfterPartialPlayback : .keepTextFallback
-  }
-}
-
-enum RealtimeHubToolFailureKind: String, Equatable {
-  case backendUnauthorized = "backend_unauthorized"
-  case backendRateLimited = "backend_rate_limited"
-  case backendClientRejected = "backend_client_rejected"
-  case backendServerError = "backend_server_error"
-  case backendTransport = "backend_transport"
-  case responseDecode = "response_decode"
-  case providerCredential = "provider_credential"
-  case toolExecution = "tool_execution"
-
-  static func classify(_ error: Error) -> RealtimeHubToolFailureKind {
-    if error is DecodingError { return .responseDecode }
-    if let apiError = error as? APIError {
-      switch apiError {
-      case .unauthorized:
-        return .backendUnauthorized
-      case .syncRateLimited:
-        return .backendRateLimited
-      case .invalidResponse, .decodingError:
-        return .responseDecode
-      case .httpError(let statusCode, _):
-        switch statusCode {
-        case 401, 403:
-          return .backendUnauthorized
-        case 408, 425, 429:
-          return .backendRateLimited
-        case 400..<500:
-          return .backendClientRejected
-        case 500..<600:
-          return .backendServerError
-        default:
-          return .backendTransport
-        }
-      case .unsupportedTierScopedBulkMutation, .syncUploadRejected:
-        return .backendClientRejected
-      }
-    }
-    if let credentialError = error as? CredentialHealthError {
-      switch credentialError.failureClass {
-      case .requiresLogin, .backendUnauthorized:
-        return .backendUnauthorized
-      case .paywalled, .byokEnrollmentMismatch, .providerAuthFailed, .providerQuotaExceeded:
-        return .providerCredential
-      case .backendTransient:
-        return .backendServerError
-      case .providerTransient, .providerPolicyClose, .unknown:
-        return .backendTransport
-      }
-    }
-    let nsError = error as NSError
-    if nsError.domain == NSURLErrorDomain { return .backendTransport }
-    return .toolExecution
-  }
-
-  var userFacingReason: String {
-    switch self {
-    case .backendUnauthorized:
-      return "Sign-in or account access needs attention."
-    case .backendRateLimited:
-      return "The service is rate limited; try again shortly."
-    case .backendClientRejected:
-      return "The request was rejected."
-    case .backendServerError:
-      return "The backend is temporarily unavailable."
-    case .backendTransport:
-      return "The network request failed."
-    case .responseDecode:
-      return "The response could not be read."
-    case .providerCredential:
-      return "The provider credential needs attention."
-    case .toolExecution:
-      return "The tool failed while running."
-    }
-  }
-}
-
-struct RealtimeHubToolFailure: Equatable {
-  let kind: RealtimeHubToolFailureKind
-
-  static func classify(_ error: Error) -> RealtimeHubToolFailure {
-    RealtimeHubToolFailure(kind: RealtimeHubToolFailureKind.classify(error))
-  }
-
-  func userFacingOutput(base: String) -> String {
-    "\(base) \(kind.userFacingReason)"
-  }
-}
-
-/// Physical PCM held while a replacement socket authenticates. Logical turn
-/// state remains in `VoiceTurnCoordinator`.
-struct RealtimeReplacementAudioBuffer {
-  static let maxBufferedAudioBytes = 3_840_000  // 120 s @ 16 kHz s16le
-
-  let turnID: VoiceTurnID
-  let responseID: VoiceResponseID
-  let identity: VoiceEffectIdentity
-  private(set) var audioBuffer: [Data] = []
-  private(set) var bufferedAudioBytes = 0
-
-  @discardableResult
-  mutating func appendAudio(_ pcm16k: Data) -> Bool {
-    let remaining = Self.maxBufferedAudioBytes - bufferedAudioBytes
-    guard remaining > 0 else { return false }
-    let accepted = pcm16k.count <= remaining ? pcm16k : Data(pcm16k.prefix(remaining))
-    guard !accepted.isEmpty else { return false }
-    audioBuffer.append(accepted)
-    bufferedAudioBytes += accepted.count
-    return accepted.count == pcm16k.count
-  }
-}
-
-/// Captures one PTT turn while a non-barge-in realtime session is reconnecting
-/// or its kernel context is being refreshed. Keeping it separate from the
-/// barge-in buffer prevents a fresh turn from being accidentally coalesced with
-/// a replaced response's input.
-struct RealtimeReconnectAudioBuffer {
-  static let maxBufferedAudioBytes = 3_840_000  // 120 s @ 16 kHz s16le
-
-  let turnID: VoiceTurnID
-  let responseID: VoiceResponseID
-  let identity: VoiceEffectIdentity
-  let interrupting: Bool
-  /// Opaque canonical identity that this logical input must observe. Audio may
-  /// not leave this buffer until the physical provider binding carries it.
-  private(set) var requiredContextFreshnessIdentity: String? = nil
-  private(set) var audioBuffer: [Data] = []
-  private(set) var bufferedAudioBytes = 0
-
-  @discardableResult
-  mutating func appendAudio(_ pcm16k: Data) -> Bool {
-    let remaining = Self.maxBufferedAudioBytes - bufferedAudioBytes
-    guard remaining > 0 else { return false }
-    let accepted = pcm16k.count <= remaining ? pcm16k : Data(pcm16k.prefix(remaining))
-    guard !accepted.isEmpty else { return false }
-    audioBuffer.append(accepted)
-    bufferedAudioBytes += accepted.count
-    return accepted.count == pcm16k.count
-  }
-
-  mutating func bindRequiredContextFreshnessIdentity(_ identity: String) -> Bool {
-    guard !identity.isEmpty else { return false }
-    if let existing = requiredContextFreshnessIdentity, existing != identity {
-      return false
-    }
-    requiredContextFreshnessIdentity = identity
-    return true
-  }
-}
-
-/// A physical socket replacement must not invalidate the logical response
-/// identity carried by buffered PTT input. The fresh session replays that
-/// buffer with this response ID, and its callbacks must remain attributable to
-/// the same reducer-owned turn. Ordinary teardown deliberately clears it.
-enum RealtimeHubReconnectIdentityPolicy {
-  static func responseIDAfterSessionDetach(
-    preservingReconnectAudio: Bool,
-    pendingReconnect: RealtimeReconnectAudioBuffer?
-  ) -> VoiceResponseID? {
-    guard preservingReconnectAudio, let pendingReconnect else { return nil }
-    return pendingReconnect.responseID
-  }
-
-  /// Production boundary that admits/fences provider session callbacks before
-  /// they become reducer events. Live-session object match is required so a
-  /// replaced socket cannot speak for the current hub; owner currency matches
-  /// `RealtimeHubOwnerFence.canReuseWarmSession`.
-  static func admitsSessionCallback(
-    isLiveSessionObject: Bool,
-    sessionOwnerIsCurrent: Bool
-  ) -> Bool {
-    isLiveSessionObject && sessionOwnerIsCurrent
-  }
-
-  /// After a successful reconnect handshake, only the newly admitted session ID
-  /// may complete the reducer path; a late callback carrying the superseded ID
-  /// is fenced.
-  static func admitsReconnectedSessionID(
-    callbackSessionID: VoiceSessionID,
-    liveSessionID: VoiceSessionID
-  ) -> Bool {
-    callbackSessionID == liveSessionID
-  }
-}
-
-struct InterruptedTurnPayload: Equatable {
-  let ownerID: String
-  let userText: String
-  let assistantText: String
-  let idempotencyKey: String
-  /// A successful `spawn_agent` call has already committed this turn's exchange
-  /// to the kernel. Capture that authority before a barge-in clears transport state.
-  let acceptedSpawnOwnerID: String?
-
-  init(
-    ownerID: String,
-    userText: String,
-    assistantText: String,
-    idempotencyKey: String,
-    acceptedSpawnOwnerID: String? = nil
-  ) {
-    self.ownerID = ownerID
-    self.userText = userText
-    self.assistantText = assistantText
-    self.idempotencyKey = idempotencyKey
-    self.acceptedSpawnOwnerID = acceptedSpawnOwnerID
-  }
-
-  /// User-visible chat text for a PTT-barged reply: keep streamed partial text only.
-  static func visibleAssistantText(partialAssistantText: String) -> String {
-    partialAssistantText.trimmingCharacters(in: .whitespacesAndNewlines)
-  }
-}
-
-/// Resolves and records the visible portion of a provider-failed turn before
-/// terminal cleanup clears its transcript. The provider callback invokes this
-/// policy before sending the terminal reducer event, which makes the ordering
-/// deterministic and directly testable without a live socket.
-enum RealtimeProviderFailureContinuity {
-  /// Installs the journal fence before provider teardown can terminalize the
-  /// turn. Transcript resolution may suspend, but the next context refresh can
-  /// already observe and await this exact continuity obligation.
-  @MainActor
-  static func registerCapturedTurn(
-    in ledger: RealtimeTurnPersistenceLedger,
-    continuityKey: String,
-    capturedTurnTask: Task<InterruptedTurnPayload?, Never>,
-    record: @escaping @MainActor (InterruptedTurnPayload) async -> Bool
-  ) -> Task<Bool, Never> {
-    ledger.enqueue(continuityKey: continuityKey, retainingReceipt: true) {
-      await persistCapturedTurn(
-        resolve: { await capturedTurnTask.value },
-        record: record)
-    }
-  }
-
-  static func persistCapturedTurn(
-    resolve: () async -> InterruptedTurnPayload?,
-    record: (InterruptedTurnPayload) async -> Bool
-  ) async -> Bool {
-    guard let interruptedTurn = await resolve() else { return true }
-    return await record(interruptedTurn)
-  }
-}
-
-struct RealtimeHubTranscriptResolution: Equatable {
-  let userText: String
-  let providerLanguage: String?
-  let localTranscript: String?
-  let localLanguage: String?
-  let usedLocalTranscript: Bool
-}
-
-/// Kernel-authored proof that a successful realtime `spawn_agent` invocation
-/// already materialized the producing voice exchange in the canonical journal.
-/// The receipt is accepted only when both stable turn IDs match the local
-/// continuity-key derivation; provider text can never substitute for it.
-struct RealtimeSpawnJournalReceipt: Equatable {
-  /// Canonical child identities returned alongside the kernel-owned journal
-  /// receipt. These let the realtime surface create the same immediate pill
-  /// projection as typed chat, rather than waiting for a later reconciliation
-  /// pass to discover an already-accepted background run.
-  struct PillProjection: Equatable {
-    let pillID: UUID
-    let sessionID: String
-    let runID: String
-    let attemptID: String?
-    let provider: String?
-    let title: String
-    let objective: String
-  }
-
-  let continuityKey: String
-  let userTurnID: String
-  let assistantTurnID: String
-  let assistantText: String
-  let pillProjection: PillProjection?
-
-  static func parse(
-    output: String,
-    expectedContinuityKey: String
-  ) -> RealtimeSpawnJournalReceipt? {
-    guard !expectedContinuityKey.isEmpty,
-      let data = output.data(using: .utf8),
-      let payload = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-      payload["schemaVersion"] as? Int == 1,
-      payload["ok"] as? Bool == true,
-      let raw = payload["journalReceipt"] as? [String: Any],
-      raw["accepted"] as? Bool == true,
-      let continuityKey = raw["continuityKey"] as? String,
-      continuityKey == expectedContinuityKey,
-      let userTurnID = raw["userTurnId"] as? String,
-      userTurnID
-        == KernelTurnProjection.stableTurnID(
-          continuityKey: continuityKey, role: "user"),
-      let assistantTurnID = raw["assistantTurnId"] as? String,
-      assistantTurnID
-        == KernelTurnProjection.stableTurnID(
-          continuityKey: continuityKey, role: "assistant"),
-      let rawAssistantText = raw["assistantText"] as? String,
-      let child = canonicalChild(in: payload)
-    else { return nil }
-    let assistantText = rawAssistantText.trimmingCharacters(in: .whitespacesAndNewlines)
-    guard !assistantText.isEmpty, assistantText.utf8.count <= 4_096 else { return nil }
-    return RealtimeSpawnJournalReceipt(
-      continuityKey: continuityKey,
-      userTurnID: userTurnID,
-      assistantTurnID: assistantTurnID,
-      assistantText: assistantText,
-      pillProjection: pillProjection(in: child))
-  }
-
-  /// The kernel attaches the journal receipt only after the agent-control
-  /// invocation has accepted its child run. Parse the child from that same
-  /// result; never derive identities from the provider's tool arguments.
-  /// The external surface returns one compact semantic child. The provider
-  /// mirror must describe that exact child and carry the same digest; accepting
-  /// only a journal receipt would let a missing/failed launch masquerade as a
-  /// successful voice delegation.
-  private static func canonicalChild(in payload: [String: Any]) -> [String: Any]? {
-    guard let child = payload["child"] as? [String: Any],
-      let providerResult = payload["providerResult"] as? [String: Any],
-      providerResult["schemaVersion"] as? Int == 1,
-      providerResult["ok"] is Bool,
-      let providerChild = providerResult["child"] as? [String: Any],
-      let digest = text(payload["semanticDigest"]),
-      text(providerResult["semanticDigest"]) == digest,
-      let lifecycle = child["lifecycle"] as? [String: Any],
-      let sessionID = text(child["sessionId"]),
-      let runID = text(child["runId"]),
-      let attemptID = text(child["attemptId"]),
-      text(child["title"]) != nil,
-      text(child["objective"]) != nil,
-      let provider = text(child["provider"]),
-      let state = text(lifecycle["state"]),
-      let attemptState = text(lifecycle["attemptState"]),
-      let adapterID = text(lifecycle["adapterId"]),
-      lifecycle["revision"] as? Int != nil,
-      lifecycle["updatedAtMs"] as? NSNumber != nil,
-      text(providerChild["sessionId"]) == sessionID,
-      text(providerChild["runId"]) == runID,
-      text(providerChild["attemptId"]) == attemptID,
-      text(providerChild["state"]) == state,
-      text(providerChild["attemptState"]) == attemptState,
-      text(providerChild["adapterId"]) == adapterID,
-      text(providerResult["code"]) != nil,
-      text(providerResult["message"]) != nil,
-      provider == adapterID
-    else { return nil }
-    return child
-  }
-
-  private static func text(_ value: Any?) -> String? {
-    guard let raw = value as? String else { return nil }
-    let trimmed = raw.trimmingCharacters(in: .whitespacesAndNewlines)
-    return trimmed.isEmpty ? nil : trimmed
-  }
-
-  private static func pillProjection(in child: [String: Any]) -> PillProjection? {
-    guard let pillID = text(child["pillId"]).flatMap(UUID.init(uuidString:)),
-      let sessionID = text(child["sessionId"]),
-      let runID = text(child["runId"])
-    else { return nil }
-    return PillProjection(
-      pillID: pillID,
-      sessionID: sessionID,
-      runID: runID,
-      attemptID: text(child["attemptId"]),
-      provider: text(child["provider"]),
-      title: text(child["title"]) ?? "Background agent",
-      objective: text(child["objective"]) ?? "Background agent")
-  }
-}
-
-/// A realtime provider may receive a transport-level success for a control
-/// tool whose semantic result is still a rejection (`{"ok":false,...}`).  A
-/// spawn is accepted only when the kernel returns its canonical journal
-/// receipt.  Keeping this distinction typed prevents the voice parent from
-/// being persisted as successful when no child agent was actually created.
-enum RealtimeSpawnAgentToolOutcome: Equatable {
-  case accepted(RealtimeSpawnJournalReceipt)
-  case setupNeeded(AgentPillsManager.DirectedProvider)
-  case rejected
-
-  static func classify(output: String, expectedContinuityKey: String) -> Self {
-    if let receipt = RealtimeSpawnJournalReceipt.parse(
-      output: output,
-      expectedContinuityKey: expectedContinuityKey)
-    {
-      return .accepted(receipt)
-    }
-    guard
-      let data = output.data(using: .utf8),
-      let payload = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-      let error = payload["error"] as? [String: Any],
-      error["code"] as? String == "provider_setup_needed",
-      let rawProvider = error["provider"] as? String,
-      let provider = AgentPillsManager.DirectedProvider(rawValue: rawProvider)
-    else {
-      return .rejected
-    }
-    return .setupNeeded(provider)
-  }
-}
 
 #if DEBUG
 /// Deterministic provider decisions for the hermetic desktop profile. This type
@@ -1084,426 +73,6 @@ struct RealtimeLocalProfileTurnPlan: Equatable {
   }
 }
 #endif
-
-enum RealtimeHubTranscriptPolicy {
-  static func resolve(
-    providerText: String,
-    preferredLanguages: [String],
-    localTranscript: String?,
-    localLanguage: String?
-  ) -> RealtimeHubTranscriptResolution {
-    let provider = providerText.trimmingCharacters(in: .whitespacesAndNewlines)
-    let local = localTranscript?.trimmingCharacters(in: .whitespacesAndNewlines)
-    let providerLanguage =
-      provider.isEmpty
-      ? nil : PTTLanguageIdentifier.dominantLanguage(of: provider, hints: [])
-    let providerMismatches =
-      !provider.isEmpty && !preferredLanguages.isEmpty
-      && (providerLanguage.map { !preferredLanguages.contains($0) } ?? false)
-    let localMatchesPreference =
-      preferredLanguages.isEmpty
-      || localLanguage.map(preferredLanguages.contains) == true
-    let shouldUseLocal =
-      (provider.isEmpty || providerMismatches)
-      && local?.isEmpty == false && localMatchesPreference
-
-    return RealtimeHubTranscriptResolution(
-      userText: shouldUseLocal ? local! : provider,
-      providerLanguage: providerLanguage,
-      localTranscript: local,
-      localLanguage: localLanguage,
-      usedLocalTranscript: shouldUseLocal)
-  }
-}
-
-enum RealtimeHubEventOwnership {
-  static func accepts(
-    _ identity: RealtimeHubEventIdentity?,
-    activeTurnID: VoiceTurnID?,
-    activeResponseID: VoiceResponseID?
-  ) -> Bool {
-    guard let identity else { return false }
-    return identity.turnID == activeTurnID && identity.responseID == activeResponseID
-  }
-}
-
-enum VoiceAudioIngressOwnership {
-  static func accepts(
-    turnID: VoiceTurnID,
-    activeTurnID: VoiceTurnID?,
-    capturingInput: Bool
-  ) -> Bool {
-    turnID == activeTurnID && capturingInput
-  }
-}
-
-enum RealtimeHubErrorOwnership {
-  static func owns(
-    route: VoiceTurnRoute?,
-    activeSessionID: VoiceSessionID?
-  ) -> Bool {
-    guard case .hub(let expectedSessionID) = route else { return false }
-    return expectedSessionID == nil || expectedSessionID == activeSessionID
-  }
-}
-
-enum RealtimeHubBargeInAction: Equatable {
-  case none
-  case stopPlaybackTail
-  case cancelInSession
-  case replaceSession
-
-  static func decide(
-    providerResponseInFlight: Bool,
-    playbackActive: Bool,
-    strategy: RealtimeHubBargeInStrategy
-  ) -> RealtimeHubBargeInAction {
-    if providerResponseInFlight {
-      return strategy == .freshSession ? .replaceSession : .cancelInSession
-    }
-    return playbackActive ? .stopPlaybackTail : .none
-  }
-}
-
-enum RealtimeInputPreparationResult: Equatable {
-  case accepted
-  case rejected
-}
-
-enum RealtimeInputAdmissionDecision: Equatable {
-  case admit
-  case rejectSupersededTurn
-  case rejectMissingContextIdentity
-  case rejectStaleProviderContext
-}
-
-/// Pure fail-closed oracle used by production replay and deterministic race
-/// tests. Physical transport readiness alone can never return `.admit`.
-enum RealtimeInputAdmissionPolicy {
-  static func decide(
-    pending: RealtimeReconnectAudioBuffer,
-    activeTurnID: VoiceTurnID?,
-    sessionContextFreshnessIdentity: String
-  ) -> RealtimeInputAdmissionDecision {
-    guard pending.turnID == activeTurnID else { return .rejectSupersededTurn }
-    guard let required = pending.requiredContextFreshnessIdentity, !required.isEmpty else {
-      return .rejectMissingContextIdentity
-    }
-    guard required == sessionContextFreshnessIdentity else {
-      return .rejectStaleProviderContext
-    }
-    return .admit
-  }
-}
-
-enum RealtimeProviderTurnDoneDisposition: Equatable {
-  case awaitToolContinuation
-  case finalizeLogicalTurn
-
-  static func decide(
-    pendingToolCount: Int,
-    postToolContinuationRequired: Bool
-  ) -> Self {
-    pendingToolCount > 0 || postToolContinuationRequired
-      ? .awaitToolContinuation
-      : .finalizeLogicalTurn
-  }
-}
-
-struct RealtimeAuthorizedToolInvocation {
-  let invocationID: String
-  let binding: ExternalSurfaceRunBinding
-  let turnID: VoiceTurnID
-  let callID: VoiceToolCallID
-  let effectIdentity: VoiceEffectIdentity
-  let canonicalToolName: String
-  let inputHash: String
-  let sourceObjectID: ObjectIdentifier
-  let turnEpoch: Int
-}
-
-enum RealtimeAuthorizedToolOwnership {
-  static func accepts(
-    command: AuthorizedToolExecution,
-    invocation: RealtimeAuthorizedToolInvocation,
-    activeTurnID: VoiceTurnID?,
-    activeToolIdentity: VoiceEffectIdentity?,
-    activeSourceObjectID: ObjectIdentifier?,
-    currentTurnEpoch: Int
-  ) -> Bool {
-    command.executor == .realtimeHub
-      && command.surfaceKind == "realtime_voice"
-      && command.invocationID == invocation.invocationID
-      && command.ownerID == invocation.binding.ownerID
-      && command.sessionID == invocation.binding.sessionID
-      && command.runID == invocation.binding.runID
-      && command.attemptID == invocation.binding.attemptID
-      && command.canonicalToolName == invocation.canonicalToolName
-      && command.inputHash == invocation.inputHash
-      && activeTurnID == invocation.turnID
-      && activeToolIdentity == invocation.effectIdentity
-      && activeSourceObjectID == invocation.sourceObjectID
-      && currentTurnEpoch == invocation.turnEpoch
-  }
-}
-
-enum RealtimeExternalRunTerminalPolicy {
-  static func status(for reason: VoiceTurnTerminalReason) -> ExternalSurfaceRunTerminalStatus {
-    switch reason {
-    case .success:
-      return .completed
-    case .tooShort, .silentRejected, .cancelled, .ownerChanged, .interruptedByBargeIn,
-      .explicitInterrupt, .cleanup:
-      return .cancelled
-    case .permissionDenied, .captureFailed, .transcriptionFailed, .providerFailed,
-      .providerNoResponse, .hubWarmTimeout, .deferredCommitTimeout,
-      .bargeInReplacementTimeout, .toolTimeout, .playbackFailed, .journalFailed:
-      return .failed
-    }
-  }
-}
-
-enum RealtimeExternalRunPromptPolicy {
-  enum Source: Equatable {
-    case finalizedTranscript
-    case partialTranscript
-    case authorizedToolFallback
-  }
-
-  struct Selection: Equatable {
-    let prompt: String
-    let source: Source
-  }
-
-  /// A realtime tool call has already passed the current-session, current-turn,
-  /// and reducer-owned capability checks before this policy is reached. Gemini
-  /// can send that call before (or instead of) a final input transcript and wait
-  /// for the result, so waiting for a final transcript here creates a circular
-  /// wait. A partial transcript is still the user's actual request and retains
-  /// target-sensitive policy context (for example, rejecting a request about
-  /// another app). Permission tools fail closed if no transcript exists at all:
-  /// their type alone must not fabricate user authority or discard an external
-  /// app target.
-  static func promptForAuthorizedTool(
-    transcript: String,
-    isFinal: Bool,
-    toolName: String = "",
-    arguments: [String: Any] = [:]
-  ) -> Selection? {
-    let prompt = transcript.trimmingCharacters(in: .whitespacesAndNewlines)
-    if !prompt.isEmpty {
-      return Selection(
-        prompt: prompt,
-        source: isFinal ? .finalizedTranscript : .partialTranscript)
-    }
-    guard !isPermissionTool(toolName, arguments: arguments) else { return nil }
-    return Selection(
-      prompt: "A realtime voice provider has already authorized one tool invocation for this active turn. Execute only that separately authorized invocation. Do not infer, expand, or perform any additional user request.",
-      source: .authorizedToolFallback)
-  }
-
-  private static func isPermissionTool(_ toolName: String, arguments: [String: Any]) -> Bool {
-    RealtimePermissionToolIdentityPolicy.contains(toolName)
-  }
-}
-
-private enum RealtimePermissionToolIdentityPolicy {
-  private static let names: Set<String> = [
-    GeneratedSwiftTool.requestPermission.rawValue,
-    GeneratedSwiftTool.checkPermissionStatus.rawValue,
-  ]
-
-  static func contains(_ toolName: String) -> Bool {
-    names.contains(toolName.trimmingCharacters(in: .whitespacesAndNewlines).lowercased())
-  }
-}
-
-/// `ptt_test_turn` injects text as the authoritative user request while still
-/// exercising the real provider, reducer, kernel, and native-tool path. Live
-/// providers may emit a bogus input transcription for the synthetic silence
-/// used to keep their activity window valid. That transport artifact must not
-/// replace the exact harness request used for authorization or persistence.
-enum RealtimeAutomationTranscriptOverridePolicy {
-  struct Selection: Equatable {
-    let text: String
-    let isFinal: Bool
-    let usedOverride: Bool
-  }
-
-  static func select(
-    providerText: String,
-    providerIsFinal: Bool,
-    forcedText: String?
-  ) -> Selection {
-    let forced = forcedText?.trimmingCharacters(in: .whitespacesAndNewlines) ?? ""
-    if !forced.isEmpty {
-      return Selection(text: forced, isFinal: true, usedOverride: true)
-    }
-    return Selection(text: providerText, isFinal: providerIsFinal, usedOverride: false)
-  }
-}
-
-@MainActor
-enum RealtimeAutomationTurnHarness {
-  /// Synthetic PTT owns a real reducer capture boundary even though it does not
-  /// start CoreAudio. Without this event, spoken fixtures longer than the
-  /// capture-start deadline terminalize before their provider commit.
-  static func begin(on coordinator: VoiceTurnCoordinator) -> VoiceTurnID {
-    let turnID = coordinator.begin(intent: .automation)
-    coordinator.send(.captureStarted(turnID: turnID, captureID: VoiceCaptureID(1)))
-    return turnID
-  }
-}
-
-/// Voice providers do not have a uniform "input transcript final" event. OpenAI
-/// eventually emits a final, while Gemini Live may only send non-final updates
-/// before it proposes a native permission tool. A permission proposal therefore
-/// collects the bounded live-transcription window, then requires the stream to
-/// be quiet rather than deriving authority from the provider's type-only tool
-/// arguments. Waiting through the complete window matters: an early quiet gap
-/// must not authorize before a later suffix can name a non-Omi app. The bounded
-/// wait keeps the provider's tool turn moving and never converts missing context
-/// into an Omi permission request.
-enum RealtimePermissionTranscriptSettlementPolicy {
-  static let quietPeriod: TimeInterval = 0.2
-  static let maximumWait: TimeInterval = 1.0
-
-  enum Decision: Equatable {
-    case execute
-    case wait(TimeInterval)
-    case reject
-  }
-
-  static func decision(
-    toolName: String,
-    transcriptIsFinal: Bool,
-    hasTranscript: Bool,
-    lastTranscriptUpdate: Date?,
-    requestStartedAt: Date,
-    now: Date
-  ) -> Decision {
-    guard isPermissionTool(toolName) else { return .execute }
-    guard !transcriptIsFinal else { return hasTranscript ? .execute : .reject }
-
-    let deadline = requestStartedAt.addingTimeInterval(maximumWait)
-    guard now >= deadline else {
-      return .wait(deadline.timeIntervalSince(now))
-    }
-    guard
-      hasTranscript,
-      let lastTranscriptUpdate,
-      now >= lastTranscriptUpdate.addingTimeInterval(quietPeriod)
-    else { return .reject }
-    return .execute
-  }
-
-  static func isPermissionTool(_ toolName: String) -> Bool {
-    RealtimePermissionToolIdentityPolicy.contains(toolName)
-  }
-}
-
-enum RealtimeVoiceContextRefreshPolicy {
-  static func requiresRefresh(
-    currentSnapshotIdentity: String,
-    sessionSnapshotIdentity: String
-  ) -> Bool {
-    currentSnapshotIdentity != sessionSnapshotIdentity
-  }
-}
-
-enum RealtimeToolTurnOwnership {
-  static func accepts(
-    turnID: VoiceTurnID,
-    identity: VoiceEffectIdentity,
-    sourceObjectID: ObjectIdentifier,
-    turnEpoch: Int,
-    activeTurnID: VoiceTurnID?,
-    activeToolIdentity: VoiceEffectIdentity?,
-    activeSourceObjectID: ObjectIdentifier?,
-    currentTurnEpoch: Int
-  ) -> Bool {
-    activeTurnID == turnID
-      && activeToolIdentity == identity
-      && activeSourceObjectID == sourceObjectID
-      && currentTurnEpoch == turnEpoch
-      && identity.generation == turnID.rawValue
-  }
-}
-
-enum RealtimeExternalToolInvocationIdentity {
-  static func make(turnID: VoiceTurnID, providerCallID: String, toolName: String) -> String {
-    let canonical = [
-      turnID.rawValue.uuidString.lowercased(),
-      "\(providerCallID.utf8.count):\(providerCallID)",
-      "\(toolName.utf8.count):\(toolName)",
-    ].joined(separator: "|")
-    let digest = SHA256.hash(data: Data(canonical.utf8))
-      .map { String(format: "%02x", $0) }
-      .joined()
-    return "voice-tool:\(digest)"
-  }
-}
-
-enum RealtimeAuthorizedInvocationReplayGate {
-  static func shouldExecute(
-    invocationID: String,
-    completedInvocationIDs: Set<String>
-  ) -> Bool {
-    !completedInvocationIDs.contains(invocationID)
-  }
-}
-
-enum RealtimeHubBargeInContinuity {
-  enum Outcome: Equatable {
-    case started
-    case interruptedTurnPersistenceFailed
-    case contextUnavailable
-    case cancelled
-  }
-
-  static let maximumContextRefreshAttempts = 8
-
-  static func prepareReplacementSession(
-    resolveInterruptedTurn: () async -> InterruptedTurnPayload?,
-    recordInterruptedTurn: (InterruptedTurnPayload) async -> Bool,
-    refreshVoiceContext: () async -> Set<String>?,
-    startReplacementSession: () -> Void
-  ) async -> Outcome {
-    let interruptedTurn = await resolveInterruptedTurn()
-    if let interruptedTurn {
-      guard await recordInterruptedTurn(interruptedTurn) else {
-        return .interruptedTurnPersistenceFailed
-      }
-    }
-    var requiredTurnIDs = Set<String>()
-    if let interruptedTurn {
-      if !interruptedTurn.userText.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty {
-        requiredTurnIDs.insert(
-          KernelTurnProjection.stableTurnID(
-            continuityKey: interruptedTurn.idempotencyKey,
-            role: "user"))
-      }
-      if !interruptedTurn.assistantText.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty {
-        requiredTurnIDs.insert(
-          KernelTurnProjection.stableTurnID(
-            continuityKey: interruptedTurn.idempotencyKey,
-            role: "assistant"))
-      }
-    }
-    for _ in 0..<maximumContextRefreshAttempts {
-      guard !Task.isCancelled else { return .cancelled }
-      guard let refreshedTurnIDs = await refreshVoiceContext(),
-        requiredTurnIDs.isSubset(of: refreshedTurnIDs)
-      else {
-        await Task.yield()
-        continue
-      }
-      startReplacementSession()
-      return .started
-    }
-    return Task.isCancelled ? .cancelled : .contextUnavailable
-  }
-}
 
 /// Keeps the response glow tied to perceived playback instead of raw PCM chunk
 /// boundaries. Realtime providers can leave short gaps between streamed audio
