+++ desktop/macos/Desktop/Sources/FloatingControlBar/RealtimeHubInputAdmission.swift
@@ -1,0 +1,102 @@
+import Foundation
+
+/// Physical PCM held while a replacement socket authenticates. Logical turn
+/// state remains in `VoiceTurnCoordinator`.
+struct RealtimeReplacementAudioBuffer {
+  static let maxBufferedAudioBytes = 3_840_000  // 120 s @ 16 kHz s16le
+
+  let turnID: VoiceTurnID
+  let responseID: VoiceResponseID
+  let identity: VoiceEffectIdentity
+  private(set) var audioBuffer: [Data] = []
+  private(set) var bufferedAudioBytes = 0
+
+  @discardableResult
+  mutating func appendAudio(_ pcm16k: Data) -> Bool {
+    let remaining = Self.maxBufferedAudioBytes - bufferedAudioBytes
+    guard remaining > 0 else { return false }
+    let accepted = pcm16k.count <= remaining ? pcm16k : Data(pcm16k.prefix(remaining))
+    guard !accepted.isEmpty else { return false }
+    audioBuffer.append(accepted)
+    bufferedAudioBytes += accepted.count
+    return accepted.count == pcm16k.count
+  }
+}
+
+/// Captures one PTT turn while a non-barge-in realtime session is reconnecting
+/// or its kernel context is being refreshed. Keeping it separate from the
+/// barge-in buffer prevents a fresh turn from being accidentally coalesced with
+/// a replaced response's input.
+struct RealtimeReconnectAudioBuffer {
+  static let maxBufferedAudioBytes = 3_840_000  // 120 s @ 16 kHz s16le
+
+  let turnID: VoiceTurnID
+  let responseID: VoiceResponseID
+  let identity: VoiceEffectIdentity
+  let interrupting: Bool
+  /// Opaque canonical identity that this logical input must observe. Audio may
+  /// not leave this buffer until the physical provider binding carries it.
+  private(set) var requiredContextFreshnessIdentity: String? = nil
+  private(set) var audioBuffer: [Data] = []
+  private(set) var bufferedAudioBytes = 0
+
+  @discardableResult
+  mutating func appendAudio(_ pcm16k: Data) -> Bool {
+    let remaining = Self.maxBufferedAudioBytes - bufferedAudioBytes
+    guard remaining > 0 else { return false }
+    let accepted = pcm16k.count <= remaining ? pcm16k : Data(pcm16k.prefix(remaining))
+    guard !accepted.isEmpty else { return false }
+    audioBuffer.append(accepted)
+    bufferedAudioBytes += accepted.count
+    return accepted.count == pcm16k.count
+  }
+
+  mutating func bindRequiredContextFreshnessIdentity(_ identity: String) -> Bool {
+    guard !identity.isEmpty else { return false }
+    if let existing = requiredContextFreshnessIdentity, existing != identity {
+      return false
+    }
+    requiredContextFreshnessIdentity = identity
+    return true
+  }
+}
+
+enum RealtimeInputPreparationResult: Equatable {
+  case accepted
+  case rejected
+}
+
+enum RealtimeInputAdmissionDecision: Equatable {
+  case admit
+  case rejectSupersededTurn
+  case rejectMissingContextIdentity
+  case rejectStaleProviderContext
+}
+
+/// Pure fail-closed oracle used by production replay and deterministic race
+/// tests. Physical transport readiness alone can never return `.admit`.
+enum RealtimeInputAdmissionPolicy {
+  static func decide(
+    pending: RealtimeReconnectAudioBuffer,
+    activeTurnID: VoiceTurnID?,
+    sessionContextFreshnessIdentity: String
+  ) -> RealtimeInputAdmissionDecision {
+    guard pending.turnID == activeTurnID else { return .rejectSupersededTurn }
+    guard let required = pending.requiredContextFreshnessIdentity, !required.isEmpty else {
+      return .rejectMissingContextIdentity
+    }
+    guard required == sessionContextFreshnessIdentity else {
+      return .rejectStaleProviderContext
+    }
+    return .admit
+  }
+}
+
+enum RealtimeVoiceContextRefreshPolicy {
+  static func requiresRefresh(
+    currentSnapshotIdentity: String,
+    sessionSnapshotIdentity: String
+  ) -> Bool {
+    currentSnapshotIdentity != sessionSnapshotIdentity
+  }
+}
