+++ desktop/macos/Desktop/Sources/FloatingControlBar/RealtimeHubSessionPolicies.swift
@@ -1,0 +1,509 @@
+import Foundation
+
+/// Safe, non-sensitive classification for realtime WebSocket teardown messages.
+///
+/// Gemini can idle-close warm sessions with WebSocket 1008 after the socket has
+/// lived for a while, and OpenAI has a known maximum-session-duration close.
+/// Both expected lifecycle paths should re-warm quietly rather than page Sentry
+/// as production errors. Fast 1008 closes are different: they usually mean
+/// provider policy/auth/config rejection and should still be reported, but with
+/// a stable category instead of raw provider text.
+enum RealtimeHubCloseCategory: String {
+  case expectedIdleTeardown = "expected_idle_teardown"
+  case expectedSessionRotation = "expected_session_rotation"
+  case providerAuthFailed = "provider_auth_failed"
+  case providerQuotaExceeded = "provider_quota_exceeded"
+  case providerPolicyCloseFast = "provider_policy_close_fast"
+  case providerTransient = "provider_transient"
+}
+
+/// The controller owns transport replacement, while the voice-turn reducer owns
+/// any active logical turn. Keeping this plan typed prevents provider-close text
+/// from leaking into lifecycle decisions outside the classifier boundary.
+enum RealtimeHubSessionRotationPlan: Equatable {
+  case rewarmIdleTransport
+  case terminateActiveTurnAndRewarm
+}
+
+enum RealtimeHubCloseClassifier {
+  static let idleTeardownThreshold: TimeInterval = 60
+
+  static func category(
+    message: String,
+    aliveFor: TimeInterval,
+    hasActiveTurn: Bool = false,
+    provider: RealtimeHubProvider = .openai
+  ) -> RealtimeHubCloseCategory? {
+    let lower = message.lowercased()
+    if provider == .openai,
+      lower.contains("your session hit the maximum duration")
+      && lower.contains("60 minutes")
+    {
+      return .expectedSessionRotation
+    }
+    guard lower.contains("websocket closed (1008)") else { return nil }
+    if CredentialHealthManager.classifyProviderClose(
+      message: message,
+      provider: provider) == .providerQuotaExceeded(provider: provider)
+    {
+      return .providerQuotaExceeded
+    }
+    if CredentialHealthManager.classifyProviderClose(
+      message: message,
+      provider: provider) == .providerAuthFailed(provider: provider, mode: .byok)
+    {
+      return .providerAuthFailed
+    }
+    if !hasActiveTurn && aliveFor >= idleTeardownThreshold { return .expectedIdleTeardown }
+    return .providerPolicyCloseFast
+  }
+
+  static func sessionRotationPlan(
+    for category: RealtimeHubCloseCategory?,
+    hasActiveTurn: Bool
+  ) -> RealtimeHubSessionRotationPlan? {
+    guard category == .expectedSessionRotation else { return nil }
+    return hasActiveTurn ? .terminateActiveTurnAndRewarm : .rewarmIdleTransport
+  }
+
+  static func isExpectedLifecycleClose(_ category: RealtimeHubCloseCategory?) -> Bool {
+    category == .expectedIdleTeardown || category == .expectedSessionRotation
+  }
+
+  static func shouldReportToSentry(_ category: RealtimeHubCloseCategory?) -> Bool {
+    !isExpectedLifecycleClose(category)
+  }
+}
+
+enum RealtimeHubCommitResult: Equatable {
+  case accepted
+  case deferredForReplacement
+  case deferredForReconnect
+  case alreadyOwned
+  case rejectedNoSession
+}
+
+enum RealtimeHubCommitOwnershipPolicy {
+  static func isAlreadyOwned(turn: VoiceTurn?, requestedTurnID: VoiceTurnID) -> Bool {
+    guard let turn, turn.id == requestedTurnID, turn.phase == .awaitingResponse,
+      turn.hubCommitPending
+    else { return false }
+    if case .hub = turn.route { return true }
+    return false
+  }
+}
+
+struct RealtimeProviderToolResult: Equatable {
+  let output: String
+  let originalByteCount: Int
+  let wasOversized: Bool
+}
+
+enum RealtimeProviderToolResultPolicy {
+  static let maximumByteCount = 48 * 1024
+
+  static func prepare(name: String, output: String) -> RealtimeProviderToolResult {
+    let originalByteCount = output.utf8.count
+    if name == HubTool.spawnAgent.rawValue,
+      let payload = try? JSONSerialization.jsonObject(with: Data(output.utf8)) as? [String: Any],
+      let providerResult = payload["providerResult"] as? [String: Any],
+      let data = try? JSONSerialization.data(withJSONObject: providerResult, options: [.sortedKeys]),
+      let compact = String(data: data, encoding: .utf8)
+    {
+      return RealtimeProviderToolResult(
+        output: compact,
+        originalByteCount: originalByteCount,
+        wasOversized: false)
+    }
+    guard originalByteCount <= maximumByteCount else {
+      let payload: [String: Any] = [
+        "ok": false,
+        "error": [
+          "code": "tool_result_too_large",
+          "message": "The tool returned too much detail. Retry with narrower filters.",
+          "tool": String(name.prefix(128)),
+          "originalBytes": originalByteCount,
+        ],
+      ]
+      let data = try? JSONSerialization.data(withJSONObject: payload, options: [.sortedKeys])
+      let bounded = data.flatMap { String(data: $0, encoding: .utf8) }
+        ?? #"{"ok":false,"error":{"code":"tool_result_too_large"}}"#
+      return RealtimeProviderToolResult(
+        output: bounded,
+        originalByteCount: originalByteCount,
+        wasOversized: true)
+    }
+    return RealtimeProviderToolResult(
+      output: output,
+      originalByteCount: originalByteCount,
+      wasOversized: false)
+  }
+}
+
+struct RealtimeHubLifecycleSnapshot: Equatable {
+  let capturingInput: Bool
+  let providerActive: Bool
+  let playbackActive: Bool
+  let pendingToolCount: Int
+  let coordinatorTurnActive: Bool
+  let minting: Bool
+}
+
+enum RealtimeHubLifecyclePolicy {
+  static func canReplaceSession(_ snapshot: RealtimeHubLifecycleSnapshot) -> Bool {
+    !snapshot.capturingInput
+      && !snapshot.providerActive
+      && !snapshot.playbackActive
+      && snapshot.pendingToolCount == 0
+      && !snapshot.coordinatorTurnActive
+      && !snapshot.minting
+  }
+
+  static func canStartGeneralWarmSession(replacementPending: Bool) -> Bool {
+    !replacementPending
+  }
+
+  static func shouldResumeCanceledTurnRefresh(
+    fenceTurnID: VoiceTurnID?,
+    terminalTurnID: VoiceTurnID
+  ) -> Bool {
+    fenceTurnID != terminalTurnID
+  }
+}
+
+/// Immutable account identity attached to a realtime socket, its context, and
+/// every token mint that may create or replace it. `signedOut` is a real scope,
+/// distinct from the absence of a session, so a later sign-in cannot inherit a
+/// socket that was warmed from signed-out/BYOK state.
+enum RealtimeHubOwnerScope: Equatable, Sendable {
+  case authenticated(String)
+  case signedOut
+
+  static func capture(currentOwnerID: String?) -> RealtimeHubOwnerScope {
+    guard let ownerID = currentOwnerID?.trimmingCharacters(in: .whitespacesAndNewlines),
+      !ownerID.isEmpty
+    else { return .signedOut }
+    return .authenticated(ownerID)
+  }
+
+  var authenticatedOwnerID: String? {
+    guard case .authenticated(let ownerID) = self else { return nil }
+    return ownerID
+  }
+
+  func isCurrent(currentOwnerID: String?) -> Bool {
+    self == Self.capture(currentOwnerID: currentOwnerID)
+  }
+}
+
+#if DEBUG
+struct RealtimeHubOwnerBoundarySnapshot: Equatable {
+  let hasPhysicalSession: Bool
+  let physicalOwnerID: String?
+  let prefetchedOwnerID: String?
+  let prefetchedContextIsEmpty: Bool
+  let hasPendingOwnerWork: Bool
+  let hubConnected: Bool
+  let turnAudioByteCount: Int
+}
+
+/// Exact non-production capability for the hermetic local-profile transport.
+/// A process-wide "test mode" boolean is not enough: the authority is bound to
+/// one physical session and one immutable owner scope, so a replaced socket or
+/// an owner transition cannot inherit the provider-warm bypass.
+struct RealtimeLocalProfileTransportAuthority: Equatable {
+  let sourceID: ObjectIdentifier
+  let ownerScope: RealtimeHubOwnerScope
+  let authorizationSnapshot: RuntimeOwnerAuthorizationSnapshot
+
+  func accepts(
+    sourceID candidateSourceID: ObjectIdentifier?,
+    currentOwnerID: String?,
+    localProfileEnabled: Bool,
+    authorizationIsCurrent: Bool
+  ) -> Bool {
+    localProfileEnabled
+      && candidateSourceID == sourceID
+      && ownerScope.isCurrent(currentOwnerID: currentOwnerID)
+      && authorizationIsCurrent
+  }
+}
+#endif
+
+/// Shared owner policy used by warm reuse, delayed mint completion, and
+/// barge-in replacement. Keeping these three paths on one decision surface
+/// prevents a future reconnect path from accidentally weakening the fence.
+enum RealtimeHubOwnerFence {
+  static func canReuseWarmSession(
+    sessionOwner: RealtimeHubOwnerScope?,
+    currentOwnerID: String?
+  ) -> Bool {
+    guard let sessionOwner else { return false }
+    return sessionOwner.isCurrent(currentOwnerID: currentOwnerID)
+  }
+
+  static func acceptsMintCompletion(
+    mintOwner: RealtimeHubOwnerScope,
+    currentOwnerID: String?
+  ) -> Bool {
+    mintOwner.isCurrent(currentOwnerID: currentOwnerID)
+  }
+
+  static func acceptsBargeInReplacement(
+    sessionOwner: RealtimeHubOwnerScope?,
+    replacementOwner: RealtimeHubOwnerScope?,
+    currentOwnerID: String?
+  ) -> Bool {
+    guard let sessionOwner, let replacementOwner, sessionOwner == replacementOwner else {
+      return false
+    }
+    return replacementOwner.isCurrent(currentOwnerID: currentOwnerID)
+  }
+}
+
+enum RealtimeNativeAudioScheduleFailureAction: Equatable {
+  case keepTextFallback
+  case failTurnAfterPartialPlayback
+
+  static func decide(playbackAlreadyStarted: Bool) -> RealtimeNativeAudioScheduleFailureAction {
+    playbackAlreadyStarted ? .failTurnAfterPartialPlayback : .keepTextFallback
+  }
+}
+
+enum RealtimeHubToolFailureKind: String, Equatable {
+  case backendUnauthorized = "backend_unauthorized"
+  case backendRateLimited = "backend_rate_limited"
+  case backendClientRejected = "backend_client_rejected"
+  case backendServerError = "backend_server_error"
+  case backendTransport = "backend_transport"
+  case responseDecode = "response_decode"
+  case providerCredential = "provider_credential"
+  case toolExecution = "tool_execution"
+
+  static func classify(_ error: Error) -> RealtimeHubToolFailureKind {
+    if error is DecodingError { return .responseDecode }
+    if let apiError = error as? APIError {
+      switch apiError {
+      case .unauthorized:
+        return .backendUnauthorized
+      case .syncRateLimited:
+        return .backendRateLimited
+      case .invalidResponse, .decodingError:
+        return .responseDecode
+      case .httpError(let statusCode, _):
+        switch statusCode {
+        case 401, 403:
+          return .backendUnauthorized
+        case 408, 425, 429:
+          return .backendRateLimited
+        case 400..<500:
+          return .backendClientRejected
+        case 500..<600:
+          return .backendServerError
+        default:
+          return .backendTransport
+        }
+      case .unsupportedTierScopedBulkMutation, .syncUploadRejected:
+        return .backendClientRejected
+      }
+    }
+    if let credentialError = error as? CredentialHealthError {
+      switch credentialError.failureClass {
+      case .requiresLogin, .backendUnauthorized:
+        return .backendUnauthorized
+      case .paywalled, .byokEnrollmentMismatch, .providerAuthFailed, .providerQuotaExceeded:
+        return .providerCredential
+      case .backendTransient:
+        return .backendServerError
+      case .providerTransient, .providerPolicyClose, .unknown:
+        return .backendTransport
+      }
+    }
+    let nsError = error as NSError
+    if nsError.domain == NSURLErrorDomain { return .backendTransport }
+    return .toolExecution
+  }
+
+  var userFacingReason: String {
+    switch self {
+    case .backendUnauthorized:
+      return "Sign-in or account access needs attention."
+    case .backendRateLimited:
+      return "The service is rate limited; try again shortly."
+    case .backendClientRejected:
+      return "The request was rejected."
+    case .backendServerError:
+      return "The backend is temporarily unavailable."
+    case .backendTransport:
+      return "The network request failed."
+    case .responseDecode:
+      return "The response could not be read."
+    case .providerCredential:
+      return "The provider credential needs attention."
+    case .toolExecution:
+      return "The tool failed while running."
+    }
+  }
+}
+
+struct RealtimeHubToolFailure: Equatable {
+  let kind: RealtimeHubToolFailureKind
+
+  static func classify(_ error: Error) -> RealtimeHubToolFailure {
+    RealtimeHubToolFailure(kind: RealtimeHubToolFailureKind.classify(error))
+  }
+
+  func userFacingOutput(base: String) -> String {
+    "\(base) \(kind.userFacingReason)"
+  }
+}
+
+/// A physical socket replacement must not invalidate the logical response
+/// identity carried by buffered PTT input. The fresh session replays that
+/// buffer with this response ID, and its callbacks must remain attributable to
+/// the same reducer-owned turn. Ordinary teardown deliberately clears it.
+enum RealtimeHubReconnectIdentityPolicy {
+  static func responseIDAfterSessionDetach(
+    preservingReconnectAudio: Bool,
+    pendingReconnect: RealtimeReconnectAudioBuffer?
+  ) -> VoiceResponseID? {
+    guard preservingReconnectAudio, let pendingReconnect else { return nil }
+    return pendingReconnect.responseID
+  }
+
+  /// Production boundary that admits/fences provider session callbacks before
+  /// they become reducer events. Live-session object match is required so a
+  /// replaced socket cannot speak for the current hub; owner currency matches
+  /// `RealtimeHubOwnerFence.canReuseWarmSession`.
+  static func admitsSessionCallback(
+    isLiveSessionObject: Bool,
+    sessionOwnerIsCurrent: Bool
+  ) -> Bool {
+    isLiveSessionObject && sessionOwnerIsCurrent
+  }
+
+  /// After a successful reconnect handshake, only the newly admitted session ID
+  /// may complete the reducer path; a late callback carrying the superseded ID
+  /// is fenced.
+  static func admitsReconnectedSessionID(
+    callbackSessionID: VoiceSessionID,
+    liveSessionID: VoiceSessionID
+  ) -> Bool {
+    callbackSessionID == liveSessionID
+  }
+}
+
+enum RealtimeHubEventOwnership {
+  static func accepts(
+    _ identity: RealtimeHubEventIdentity?,
+    activeTurnID: VoiceTurnID?,
+    activeResponseID: VoiceResponseID?
+  ) -> Bool {
+    guard let identity else { return false }
+    return identity.turnID == activeTurnID && identity.responseID == activeResponseID
+  }
+}
+
+enum VoiceAudioIngressOwnership {
+  static func accepts(
+    turnID: VoiceTurnID,
+    activeTurnID: VoiceTurnID?,
+    capturingInput: Bool
+  ) -> Bool {
+    turnID == activeTurnID && capturingInput
+  }
+}
+
+enum RealtimeHubErrorOwnership {
+  static func owns(
+    route: VoiceTurnRoute?,
+    activeSessionID: VoiceSessionID?
+  ) -> Bool {
+    guard case .hub(let expectedSessionID) = route else { return false }
+    return expectedSessionID == nil || expectedSessionID == activeSessionID
+  }
+}
+
+enum RealtimeHubBargeInAction: Equatable {
+  case none
+  case stopPlaybackTail
+  case cancelInSession
+  case replaceSession
+
+  static func decide(
+    providerResponseInFlight: Bool,
+    playbackActive: Bool,
+    strategy: RealtimeHubBargeInStrategy
+  ) -> RealtimeHubBargeInAction {
+    if providerResponseInFlight {
+      return strategy == .freshSession ? .replaceSession : .cancelInSession
+    }
+    return playbackActive ? .stopPlaybackTail : .none
+  }
+}
+
+enum RealtimeProviderTurnDoneDisposition: Equatable {
+  case awaitToolContinuation
+  case finalizeLogicalTurn
+
+  static func decide(
+    pendingToolCount: Int,
+    postToolContinuationRequired: Bool
+  ) -> Self {
+    pendingToolCount > 0 || postToolContinuationRequired
+      ? .awaitToolContinuation
+      : .finalizeLogicalTurn
+  }
+}
+
+enum RealtimeHubBargeInContinuity {
+  enum Outcome: Equatable {
+    case started
+    case interruptedTurnPersistenceFailed
+    case contextUnavailable
+    case cancelled
+  }
+
+  static let maximumContextRefreshAttempts = 8
+
+  static func prepareReplacementSession(
+    resolveInterruptedTurn: () async -> InterruptedTurnPayload?,
+    recordInterruptedTurn: (InterruptedTurnPayload) async -> Bool,
+    refreshVoiceContext: () async -> Set<String>?,
+    startReplacementSession: () -> Void
+  ) async -> Outcome {
+    let interruptedTurn = await resolveInterruptedTurn()
+    if let interruptedTurn {
+      guard await recordInterruptedTurn(interruptedTurn) else {
+        return .interruptedTurnPersistenceFailed
+      }
+    }
+    var requiredTurnIDs = Set<String>()
+    if let interruptedTurn {
+      if !interruptedTurn.userText.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty {
+        requiredTurnIDs.insert(
+          KernelTurnProjection.stableTurnID(
+            continuityKey: interruptedTurn.idempotencyKey,
+            role: "user"))
+      }
+      if !interruptedTurn.assistantText.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty {
+        requiredTurnIDs.insert(
+          KernelTurnProjection.stableTurnID(
+            continuityKey: interruptedTurn.idempotencyKey,
+            role: "assistant"))
+      }
+    }
+    for _ in 0..<maximumContextRefreshAttempts {
+      guard !Task.isCancelled else { return .cancelled }
+      guard let refreshedTurnIDs = await refreshVoiceContext(),
+        requiredTurnIDs.isSubset(of: refreshedTurnIDs)
+      else {
+        await Task.yield()
+        continue
+      }
+      startReplacementSession()
+      return .started
+    }
+    return Task.isCancelled ? .cancelled : .contextUnavailable
+  }
+}
